+++ Plugin/CodeRebirth/src/Content/Unlockables/CruiserCharger.cs
@@ -1,5 +1,5 @@
 using Unity.Netcode;
-/*
+
 namespace CodeRebirth.src.Content.Unlockables;
 public class CruiserCharger : Charger
 {
@@ -8,9 +8,9 @@ public class CruiserCharger : Charger
         base.OnNetworkSpawn();
         if (!IsServer) return;
         // Instantiate the CruiserGalAI prefab
-        GalAI = Instantiate(UnlockableHandler.Instance.CruiserBot.CruiserDronePrefab, ChargeTransform.position, ChargeTransform.rotation).GetComponent<CruiserGalAI>();
+        GalAI = Instantiate(UnlockableHandler.Instance.CruiserGal.CruiserGalPrefab, ChargeTransform.position, ChargeTransform.rotation).GetComponent<CruiserGalAI>();
         NetworkObject netObj = GalAI.GetComponent<NetworkObject>();
         GalAI.GalCharger = this;
         netObj.Spawn();
     }
-}*/
+}